backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yeonggwang <immutable000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
Pillow = "^8.4.0"
djangorestframework-simplejwt = "^5.0.0"
drf-yasg = "^1.20.0"
django-dotenv = "^1.4.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
